+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.12.2
 - Added `gitGrace.stageAll`, `gitGrace.unstageAll`, and `gitGrace.cleanAll` commands in favor of the built-in commands.
+- Fixed `gitGrace.blame` and `gitGrace.openWeb` commands so they use the pushed commit hash.
 
 ### 2.12.1
 - Added keybinding of `git.stageAll` command.